refactor: PhotoFrame 이미지 업로드 로직을 커스텀 훅으로 분리(#130)
